Polish pass: v1.1.0, app icon, About box, packaging, warning-clean build
- Version bumped to 1.1.0, defined once in CMake and stamped into the window
  title, the About box, and the Windows version resource (generated .rc).
- App icon: phosphor-green Lissajous .ico embedded in the exe and applied to
  the window/taskbar via WM_SETICON.
- Help > About shows version, GL renderer, build date, license.
- Escape no longer quits (one stray keypress closed the app with no confirm);
  Space no longer toggles playback while typing in a text field.
- make_release.ps1 builds the standalone exe and packages
  dist/oscilloplot-v<version>-windows-x64.zip; version read from CMakeLists.
- imgui.ini untracked (user-specific layout, already gitignored).
- Removed the vectorizer_tool target and source: it duplicated the in-app
  Image Vectorizer and doubled build time and release size.
- Warning-clean build: fixed all project C4244s (double->float via M_PI and
  std::pow promotions), the C4458 shadow in SineLUT, silenced intentional
  C4324 alignas padding, and compile vendored imgui/implot at /W0 so project
  warnings are the only warnings.

Features:
- Vectorizer: importance mask now feeds the Canny thresholds (it was computed
  and discarded); new Detail Emphasis slider controls how strongly.
- Drawing canvas: per-stroke Undo (button + Ctrl+Z when focused).
- 3D shapes: Bake Rotation freezes N animation frames into one long pattern so
  playback and WAV export actually rotate; capped at the engine's 65,536-point
  buffer with a truncation notice.

+++ src/audio/effects/kaleidoscope.cpp
@@ -24,7 +24,7 @@ void KaleidoscopeEffect::process(Pattern& pattern, float time, float deltaTime) 
     // Expand pattern to include all segments
     pattern.resize(origSize * m_segments);
 
-    float segmentAngle = 2.0f * M_PI / m_segments;
+    float segmentAngle = 2.0f * static_cast<float>(M_PI) / static_cast<float>(m_segments);
 
     for (int seg = 0; seg < m_segments; ++seg) {
         float angle = seg * segmentAngle + m_rotation;
